double=>float, 5% memory savings

+++ src/main/java/edu/cmu/cs/ark/semeval2014/lr/FVItem.java
@@ -3,8 +3,8 @@ package edu.cmu.cs.ark.semeval2014.lr;
 public class FVItem {
 	int featnum;
 	int label;
-	double value;
+	float value;
 	
-	FVItem(int f, int k, double v) { featnum=f; label=k; value=v; }
+	FVItem(int f, int k, float v) { featnum=f; label=k; value=v; }
 
 }
